refactor:Issue-25:Move code to monitoring-agent-lib
Moves all read of /proc information to the library.

Moves proc cpu to library.
Moves proc mem to library.
Removes unused code from daemon.

Models in library still have serde serialization/deserialization. This
will be removed later.

Breaking changes: None

Resolves: #25
