+++ Cargo.toml
@@ -1,4 +1,6 @@
 [workspace]
+resolver = "2"
+
 members = [ 
     "monitoring-agent-daemon",
     "monitoring-agent-daemon",
