+++ monitoring-agent-daemon/src/common/mod.rs
@@ -5,10 +5,8 @@ mod applicationerror;
 mod monitorstatus;
 pub mod configuration;
 pub mod args;
-pub mod procsdata;
 
 pub use crate::common::applicationerror::ApplicationError;
 pub use crate::common::monitorstatus::{MonitorStatus, Status};
 pub use crate::common::configuration::{Monitor, MonitorType, HttpMethod};
 pub use crate::common::args::ApplicationArguments;
-pub use crate::common::procsdata::{ProcsCpuinfo, ProcsMeminfo};
