+++ monitoring-agent-lib/Cargo.toml
@@ -4,3 +4,6 @@ version = "0.1.0"
 edition = "2021"
 
 [dependencies]
+serde = { version = "1.0.204", features = ["derive", ] }                                # For serializing/deserializing.
+serde_json = { version = "1.0.120", features = []}                                      # For serializing/deserializing.
+log = "0.4.22"                                                                          # For logging.
