+++ monitoring-agent-lib/resources/test/test_cpuinfo
