+++ monitoring-agent-lib/src/common/mod.rs
@@ -1,0 +1,6 @@
+/*
+ * Common functionality for the library.
+*/
+pub mod error;
+
+pub use self::error::CommonLibError;
