+++ monitoring-agent-lib/src/proc/mod.rs
@@ -1,0 +1,8 @@
+/* 
+ * Common structures for reading proc files.
+ */
+pub mod cpuinfo;
+pub mod meminfo;
+
+pub use crate::proc::cpuinfo::ProcsCpuinfo;
+pub use crate::proc::meminfo::ProcsMeminfo;
